requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
httpx>=0.27.0
discord.py>=2.4.0
